feat: implement CourseContents page, enhance Course retrieval with selected category, and update user authentication flow

+++ LearnStream/frontend/src/components/Navbar1.jsx
@@ -1,5 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { Avatar, Dropdown, Navbar } from "flowbite-react";
+import { useNavigate } from "react-router-dom";
+import axios from "../api/axios";
 // import {user.jpg} from "../assests/user.jpg"
 
 const Navbar1 = () => {
@@ -7,21 +9,40 @@ const Navbar1 = () => {
   const [accessToken, setAccessToken] = useState(null);
   const [avatarUrl, setAvatarUrl] = useState(null);
   const [userName, setUserName] = useState("Guest");
-
+  const navigate = useNavigate()
   // Fetch user details from localStorage when component mounts
   useEffect(() => {
     setUserId(localStorage.getItem("user_id"));
-    setAccessToken(localStorage.getItem("accesstoken"));
-    setAvatarUrl(
-      localStorage.getItem("avatarUrl") || "https://www.google.com/url?sa=i&url=https%3A%2F%2Fstock.adobe.com%2Fsearch%3Fk%3Ddefault%2Bavatar&psig=AOvVaw04t_GWwT7vB3zbrFgLQkWh&ust=1737262814023000&source=images&cd=vfe&opi=89978449&ved=0CBEQjRxqFwoTCJCxsLm-_ooDFQAAAAAdAAAAABAE" // Default placeholder
-    );
+    setAccessToken(localStorage.getItem("accessToken"));
+    // setAvatarUrl(
+    //   localStorage.getItem("avatarUrl") || "https://www.google.com/url?sa=i&url=https%3A%2F%2Fstock.adobe.com%2Fsearch%3Fk%3Ddefault%2Bavatar&psig=AOvVaw04t_GWwT7vB3zbrFgLQkWh&ust=1737262814023000&source=images&cd=vfe&opi=89978449&ved=0CBEQjRxqFwoTCJCxsLm-_ooDFQAAAAAdAAAAABAE" // Default placeholder
+    // );
     setUserName(localStorage.getItem("user_name") || "Guest");
   }, []);
 
-  const handleLogout = () => {
-    // Clear localStorage and redirect to home
-    localStorage.clear();
-    window.location.href = "/";
+  const handleLogout =async () => {
+    try {
+      const response = await axios.post(
+        `/user/${localStorage.getItem('role')}/logout`,
+        {}, // Empty body
+        {
+          headers: {
+            'Content-Type': 'application/json',
+            Authorization: `Bearer ${localStorage.getItem('accessToken')}`,
+          },
+          withCredentials: true,
+        }
+      );
+      
+      if (response){
+        localStorage.clear();
+        navigate('/');
+      }
+      else throw Error;
+    } catch (error) {
+      console.log(error)
+    }
+    
   };
 
   return (
